+++ src/backend/joanie/core/flows/order.py
@@ -164,6 +164,7 @@ class OrderFlow:
     @state.transition(
         source=[
             enums.ORDER_STATE_ASSIGNED,
+            enums.ORDER_STATE_NO_PAYMENT,
             enums.ORDER_STATE_PENDING_PAYMENT,
             enums.ORDER_STATE_FAILED_PAYMENT,
             enums.ORDER_STATE_PENDING,